+++ src/app/components/HomeSelectors/HomeSelectors.module.scss
@@ -3,15 +3,17 @@
 
 .homeSelectors {
   display: flex;
-  gap: $spacing-s;
-  font-size: 0.75rem;
+  gap: $spacing-l;
+  font-size: 1rem;
 
-  @include from-tablet {
-    gap: $spacing-l;
-    font-size: 1rem;
-  }
+  label {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: $spacing-s;
 
-  select {
-    margin-left: $spacing-s;
+    @include from-tablet {
+      flex-direction: row;
+    }
   }
 }
