+++ src/app/page.tsx
@@ -4,6 +4,7 @@ import React, { useEffect, useState } from 'react';
 import styles from '@/page.module.scss';
 import toast from 'react-hot-toast';
 import { useNewsContext } from '@/context/news.context';
+import { useRouter } from 'next/navigation';
 import NewsList from './components/NewsList/NewsList';
 import Spinner from './components/Spinner/Spinner';
 import getNews from './utils/getNews';
@@ -19,7 +20,16 @@ function Home({ searchParams }: Props) {
   const { news, loading, setNews, setLoading } = useNewsContext();
 
   const [selects, setSelects] = useState<Selects>({ language: 'en', pageSize: '9' });
+
+  const router = useRouter();
+  const totalPageNumber = Math.floor(100 / news.length); // 100 articles - API limit
   const [currentPage, setCurrentPage] = useState(1);
+
+  const handlePageChange = (value: number) => {
+    setCurrentPage(Math.max(1, Math.min(totalPageNumber, value)));
+    const query = `?q=${searchParams.q}&page=${value}`;
+    router.push(query);
+  };
 
   useEffect(() => {
     (async () => {
@@ -53,7 +63,11 @@ function Home({ searchParams }: Props) {
           <>
             <h2>Search results for: {searchParams.q}</h2>
             <NewsList news={news} />
-            <Pagination currentPage={currentPage} setCurrentPage={setCurrentPage} />
+            <Pagination
+              currentPage={currentPage}
+              totalPageNumber={totalPageNumber}
+              handlePageChange={handlePageChange}
+            />
           </>
         )
       )}
